+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## 2024-06-04
 
+### Changed
+
+- Some minor UI adjustments to the navbar and the bottom buttons.
+
 ### Fixed
 
 - YouTube thumbnail calculation now correctly removes all video query parameters to ensure the correct thumbnail is displayed.